+++ lib/config/cubit/app_cubit.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 part 'app_state.dart';
@@ -8,5 +8,15 @@ class AppCubit extends Cubit<AppState> {
 
   void changeLocale(Locale locale) {
     emit(state.copyWith(locale: locale));
+  }
+
+  void toggleThemeMode() {
+    emit(
+      state.copyWith(
+        themeMode: state.themeMode == ThemeMode.light //
+            ? ThemeMode.dark
+            : ThemeMode.light,
+      ),
+    );
   }
 }