+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ### 0.0.0
 
+- feat: Added way to switch theme mode;
 - feat: Added way to switch language;
 - feat: Application route configuration.